feat: refactor the component, change 3 columns to 2
