+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/ClosedLoan/components/CollateralRow/index.ts
@@ -1,0 +1,1 @@
+export { CollateralRow } from './CollateralRow.tsx';
