+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/ClosedLoan/index.ts
@@ -1,0 +1,1 @@
+export { ClosedLoan } from './ClosedLoan.tsx';
