+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/LoadingCard/index.ts
@@ -1,0 +1,1 @@
+export { LoadingCard } from './LoadingCard';
