+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/OpenLoan/components/CollateralCell/index.ts
@@ -1,0 +1,1 @@
+export { CollateralCell } from './CollateralCell.tsx';
