+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/OpenLoan/components/LeveragedExposure/index.ts
@@ -1,0 +1,1 @@
+export { LeveragedExposure } from './LeveragedExposure';
