+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/OpenLoan/components/NetValueCell/index.ts
@@ -1,0 +1,1 @@
+export { NetValueCell } from './NetValueCell';
