+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/OpenLoan/index.ts
@@ -1,0 +1,1 @@
+export { OpenLoan } from './OpenLoan';
