+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/components/CrossedText/index.ts
@@ -1,0 +1,1 @@
+export { CrossedText } from './CrossedText';
