+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/components/GridItem/index.ts
@@ -1,0 +1,1 @@
+export { GridItem } from './GridItem.tsx';
