+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/components/GridItemWrapper/index.ts
@@ -1,0 +1,1 @@
+export { GridItemWrapper } from './GridItemWrapper.tsx';
