+++ frontend/app/src/screens/TransactionsScreen/LoanCard/components/components/TotalDebt/index.ts
@@ -1,0 +1,1 @@
+export { TotalDebt } from './TotalDebt';
