+++ frontend/app/src/screens/TransactionsScreen/LoanCard/index.ts
@@ -1,0 +1,1 @@
+export { LoanCard } from './LoanCard';
